LPS-94616 Assigning all permissions to the root folder when they don't come in the response
When the root folder is requested the response does not include the
EffectiveBasePermissions bits. By returning 0 we were not allowing users
to create content in it when they actually can.

With this changes we allow all permissions for the root folder if they
are not included in the response thus allowing users to create content.
If sharepoint does not allow it it will fail on creation but at least
they will be able to try.