LPS-103511 Apply manual lint fixes
This is largely similar to f39ee5fa39b29aa85c6c188bffe4081094372f75 on
the master branch.

Some of them were false positives, likely caused by a bug in ESLint,
where it things variables declared with `for (const a in b)` to be
unused; in all of those cases except one I was able to switch to
`Object.keys` instead, which is better than `in` anyway because it
obviates the need to do an `hasOwnProperty()` check. The exception was a
`FormData` instance for which I am not sure whether IE will do the right
thing, so I left that one as it was and added a suppression.

The move of the `jest.mock()` call is safe because Jest transparently
hoists such calls up to the top of the module anyway via Babel at
runtime. Tests still pass in that module.