LPS-96723 Remove pointless comma operator
This pattern:

    fn(...(null, args));

looks like a mistranslation of:

    fn.apply(null, args);

ie. call function `fn` with its context (`this`) set to `null`, and
passing it `args`.

However, the comma operator has no effect. It evaluates the first
operand (`null`), discards it, then evaluates the second one (`args`)
and uses that as the result of the expression:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Comma_Operator

In other words, it can be rewritten as just:

    fn(...args);