LPS-117924 Simplify yarn workspace globs
This is the re-do of:

https://github.com/brianchandotcom/liferay-portal/pull/91892 (master PR)

for 7.2.x.

As explained in that PR, a common question we get in Slack asks why the
"liferay-npm-scripts" executable can't be found at build time, and the
problem is usually a missing glob in the Yarn workspaces list. So the
idea here is to eliminate that class of problem by making our globs much
looser by default. We were initially more conservative out of caution,
but performance seems fine with the broader globs and the build is
unchanged.

I took the "compare.js" script from that pull:

https://gist.github.com/wincent/fea9f37abe2261526cf5b9a8191296ee

And modified it for 7.2.x:

https://gist.github.com/wincent/eec43f9f7b3a34c54048fac726cb4e54

It printed:

    Common: 110, Missing: 2, Extra: 0

    Missing:

    - dxp/apps/osb/osb-commerce/osb-commerce-provisioning/osb-commerce-provisioning-theme/package.json
    - dxp/apps/osb/osb-commerce/osb-commerce-provisioning/osb-commerce-provisioning-web/package.json

If I loosen the glob to cover those (ie. with `dxp/apps/*/*/*/*`), I get
this, which is not what we want (because it picks up something in
node_modules):

    Common: 112, Missing: 0, Extra: 1

    Extra:

    - dxp/apps/portal-workflow/portal-workflow-metrics-web/node_modules/qs/package.json

So the final tweak I settled on in this commit is to use something more
specific (`dxp/apps/osb/*/*/*`), which gives us:

    Common: 112, Missing: 0, Extra: 0

ie. identical workspace definition before and after.

A `yarn install` leaves the lockfile untouched, confirming that there
are no additional side-effects to this change.